20240530_gpt接口调用/requirements.txt
fastapi==0.111.0
langid==1.1.6
pydantic==1.8.2
python-dotenv==1.0.1
requests==2.28.2
uvicorn==0.16.0
